Add support for listing and reading tags
This adds optional support to be able to list and read tags on the
instance via an additional boolean flag.

If the flag is provided we asynchronously mount a new FS (TagsFs) onto
the given mountpoint under the tags/ directory.

This requires the AWS SDK so that we can query instance tags. Valid AWS
credentials are needed for this functionality which can be specified via
command line flags, environment variables, the shared AWS credentials
file, or IAM instance profile.

Moved MetadataFs into its own package to make testing easier now that we
have an additional FS (TagsFs).

See README changes for more details.