Naive support for caching file and directory attributes
This adds a naive implementation allowing users to cache file and
directory attributes (and directory listings) via a new `cachesec`
parameter. This makes use of the caching filesystem provided by the
unionfs subpackage of go-fuse.

We default to no cache since I believe this to be the least suprising
behavior.